// integration/google-cloud-pubsub/e2e/proto/level3-extended.ts
export interface Level3ProtocolBuffer {
  /**
   * @generated from protobuf field: string field1 = 1
   */
  field1: string;
  /**
   * @generated from protobuf field: int32 field2 = 2
   */
  field2: number;
  /**
   * @generated from protobuf field: bool field3 = 3
   */
  field3: boolean;
  /**
   * @generated from protobuf field: double field4 = 4
   */
  field4: number;
  /**
   * @generated from protobuf field: example.level3.Level3ProtocolBuffer.Nested1ProtocolBuffer field5 = 5
   */
  field5?: Level3ProtocolBuffer_Nested1ProtocolBuffer;
  /**
   * @generated from protobuf field: string newField6 = 6
   */
  newField6: string;
  /**
   * @generated from protobuf field: int64 newField7 = 7
   */
  newField7: bigint;
  /**
   * @generated from protobuf field: bytes newField8 = 8
   */
  newField8: Uint8Array;
}
/**
 * @generated from protobuf message example.level3.Level3ProtocolBuffer.Nested1ProtocolBuffer
 */
export interface Level3ProtocolBuffer_Nested1ProtocolBuffer {
  /**
   * @generated from protobuf field: string nestedField1 = 1
   */
  nestedField1: string;
}
// @generated message type with reflection information, may provide speed optimized methods
class Level3ProtocolBuffer$Type extends MessageType<Level3ProtocolBuffer> {
  constructor() {
    super('example.level3.Level3ProtocolBuffer', [
      { no: 1, name: 'field1', kind: 'scalar', T: 9 /*ScalarType.STRING*/ },
      { no: 2, name: 'field2', kind: 'scalar', T: 5 /*ScalarType.INT32*/ },
      { no: 3, name: 'field3', kind: 'scalar', T: 8 /*ScalarType.BOOL*/ },
      { no: 4, name: 'field4', kind: 'scalar', T: 1 /*ScalarType.DOUBLE*/ },
      {
        no: 5,
        name: 'field5',
        kind: 'message',
        T: () => Level3ProtocolBuffer_Nested1ProtocolBuffer,
      },
      { no: 6, name: 'newField6', kind: 'scalar', T: 9 /*ScalarType.STRING*/ },
      {
        no: 7,
        name: 'newField7',
        kind: 'scalar',
        T: 3 /*ScalarType.INT64*/,
        L: 0 /*LongType.BIGINT*/,
      },
      { no: 8, name: 'newField8', kind: 'scalar', T: 12 /*ScalarType.BYTES*/ },
    ]);
  }
  create(value?: PartialMessage<Level3ProtocolBuffer>): Level3ProtocolBuffer {
    const message = globalThis.Object.create(this.messagePrototype!);
    message.field1 = '';
    message.field2 = 0;
    message.field3 = false;
    message.field4 = 0;
    message.newField6 = '';
    message.newField7 = 0n;
    message.newField8 = new Uint8Array(0);
    if (value !== undefined)
      reflectionMergePartial<Level3ProtocolBuffer>(this, message, value);
    return message;
  }
  internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: Level3ProtocolBuffer,
  ): Level3ProtocolBuffer {
    const message = target ?? this.create(),
      end = reader.pos + length;
    while (reader.pos < end) {
      const [fieldNo, wireType] = reader.tag();
      switch (fieldNo) {
        case /* string field1 */ 1:
          message.field1 = reader.string();
          break;
        case /* int32 field2 */ 2:
          message.field2 = reader.int32();
          break;
        case /* bool field3 */ 3:
          message.field3 = reader.bool();
          break;
        case /* double field4 */ 4:
          message.field4 = reader.double();
          break;
        case /* example.level3.Level3ProtocolBuffer.Nested1ProtocolBuffer field5 */ 5:
          message.field5 =
            Level3ProtocolBuffer_Nested1ProtocolBuffer.internalBinaryRead(
              reader,
              reader.uint32(),
              options,
              message.field5,
            );
          break;
        case /* string newField6 */ 6:
          message.newField6 = reader.string();
          break;
        case /* int64 newField7 */ 7:
          message.newField7 = reader.int64().toBigInt();
          break;
        case /* bytes newField8 */ 8:
          message.newField8 = reader.bytes();
          break;
        default:
          const u = options.readUnknownField;
          if (u === 'throw')
            throw new globalThis.Error(
              `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`,
            );
          const d = reader.skip(wireType);
          if (u !== false)
            (u === true ? UnknownFieldHandler.onRead : u)(
              this.typeName,
              message,
              fieldNo,
              wireType,
              d,
            );
      }
    }
    return message;
  }
  internalBinaryWrite(
    message: Level3ProtocolBuffer,
    writer: IBinaryWriter,
    options: BinaryWriteOptions,
  ): IBinaryWriter {
    /* string field1 = 1; */
    if (message.field1 !== '')
      writer.tag(1, WireType.LengthDelimited).string(message.field1);
    /* int32 field2 = 2; */
    if (message.field2 !== 0)
      writer.tag(2, WireType.Varint).int32(message.field2);
    /* bool field3 = 3; */
    if (message.field3 !== false)
      writer.tag(3, WireType.Varint).bool(message.field3);
    /* double field4 = 4; */
    if (message.field4 !== 0)
      writer.tag(4, WireType.Bit64).double(message.field4);
    /* example.level3.Level3ProtocolBuffer.Nested1ProtocolBuffer field5 = 5; */
    if (message.field5)
      Level3ProtocolBuffer_Nested1ProtocolBuffer.internalBinaryWrite(
        message.field5,
        writer.tag(5, WireType.LengthDelimited).fork(),
        options,
      ).join();
    /* string newField6 = 6; */
    if (message.newField6 !== '')
      writer.tag(6, WireType.LengthDelimited).string(message.newField6);
    /* int64 newField7 = 7; */
    if (message.newField7 !== 0n)
      writer.tag(7, WireType.Varint).int64(message.newField7);
    /* bytes newField8 = 8; */
    if (message.newField8.length)
      writer.tag(8, WireType.LengthDelimited).bytes(message.newField8);
    const u = options.writeUnknownFields;
    if (u !== false)
      (u == true ? UnknownFieldHandler.onWrite : u)(
        this.typeName,
        message,
        writer,
      );
    return writer;
  }
}
/**
 * @generated MessageType for protobuf message example.level3.Level3ProtocolBuffer
 */
export const Level3ProtocolBuffer = new Level3ProtocolBuffer$Type();
// @generated message type with reflection information, may provide speed optimized methods
class Level3ProtocolBuffer_Nested1ProtocolBuffer$Type extends MessageType<Level3ProtocolBuffer_Nested1ProtocolBuffer> {
  constructor() {
    super('example.level3.Level3ProtocolBuffer.Nested1ProtocolBuffer', [
      {
        no: 1,
        name: 'nestedField1',
        kind: 'scalar',
        T: 9 /*ScalarType.STRING*/,
      },
    ]);
  }
  create(
    value?: PartialMessage<Level3ProtocolBuffer_Nested1ProtocolBuffer>,
  ): Level3ProtocolBuffer_Nested1ProtocolBuffer {
    const message = globalThis.Object.create(this.messagePrototype!);
    message.nestedField1 = '';
    if (value !== undefined)
      reflectionMergePartial<Level3ProtocolBuffer_Nested1ProtocolBuffer>(
        this,
        message,
        value,
      );
    return message;
  }
  internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: Level3ProtocolBuffer_Nested1ProtocolBuffer,
  ): Level3ProtocolBuffer_Nested1ProtocolBuffer {
    const message = target ?? this.create(),
      end = reader.pos + length;
    while (reader.pos < end) {
      const [fieldNo, wireType] = reader.tag();
      switch (fieldNo) {
        case /* string nestedField1 */ 1:
          message.nestedField1 = reader.string();
          break;
        default:
          const u = options.readUnknownField;
          if (u === 'throw')
            throw new globalThis.Error(
              `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`,
            );
          const d = reader.skip(wireType);
          if (u !== false)
            (u === true ? UnknownFieldHandler.onRead : u)(
              this.typeName,
              message,
              fieldNo,
              wireType,
              d,
            );
      }
    }
    return message;
  }
  internalBinaryWrite(
    message: Level3ProtocolBuffer_Nested1ProtocolBuffer,
    writer: IBinaryWriter,
    options: BinaryWriteOptions,
  ): IBinaryWriter {
    /* string nestedField1 = 1; */
    if (message.nestedField1 !== '')
      writer.tag(1, WireType.LengthDelimited).string(message.nestedField1);
    const u = options.writeUnknownFields;
    if (u !== false)
      (u == true ? UnknownFieldHandler.onWrite : u)(
        this.typeName,
        message,
        writer,
      );
    return writer;
  }
}
/**
 * @generated MessageType for protobuf message example.level3.Level3ProtocolBuffer.Nested1ProtocolBuffer
 */
export const Level3ProtocolBuffer_Nested1ProtocolBuffer =
  new Level3ProtocolBuffer_Nested1ProtocolBuffer$Type();