// integration/google-cloud-pubsub/e2e/proto/test.ts
// @generated by protobuf-ts 2.11.1
// @generated from protobuf file "test.proto" (package "test", syntax proto3)
// tslint:disable
import type { BinaryWriteOptions } from '@protobuf-ts/runtime';
import type { IBinaryWriter } from '@protobuf-ts/runtime';
import type { BinaryReadOptions } from '@protobuf-ts/runtime';
import type { IBinaryReader } from '@protobuf-ts/runtime';
import type { PartialMessage } from '@protobuf-ts/runtime';
import { UnknownFieldHandler } from '@protobuf-ts/runtime';
import { WireType } from '@protobuf-ts/runtime';
import { reflectionMergePartial } from '@protobuf-ts/runtime';
import { MessageType } from '@protobuf-ts/runtime';

/**
 * @generated from protobuf message test.TestEvent
 */
export interface TestEvent {
  /**
   * @generated from protobuf field: bool isActive = 2
   */
  isActive: boolean;
  /**
   * @generated from protobuf field: string test = 1
   */
  test: string;
}

// @generated message type with reflection information, may provide speed optimized methods
class TestEvent$Type extends MessageType<TestEvent> {
  constructor() {
    super('test.TestEvent', [
      { T: 9, kind: 'scalar', name: 'test', no: 1 /*ScalarType.STRING*/ },
      { T: 8, kind: 'scalar', name: 'isActive', no: 2 /*ScalarType.BOOL*/ },
    ]);
  }

  create(value?: PartialMessage<TestEvent>): TestEvent {
    const message = globalThis.Object.create(this.messagePrototype!);

    message.test = '';
    message.isActive = false;

    if (value !== undefined)
      reflectionMergePartial<TestEvent>(this, message, value);

    return message;
  }

  internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: TestEvent,
  ): TestEvent {
    const message = target ?? this.create(),
      end = reader.pos + length;

    while (reader.pos < end) {
      const [fieldNo, wireType] = reader.tag();

      switch (fieldNo) {
        case /* string test */ 1:
          message.test = reader.string();
          break;
        case /* bool isActive */ 2:
          message.isActive = reader.bool();
          break;
        default:
          const u = options.readUnknownField;

          if (u === 'throw')
            throw new globalThis.Error(
              `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`,
            );

          const d = reader.skip(wireType);

          if (u !== false)
            (u === true ? UnknownFieldHandler.onRead : u)(
              this.typeName,
              message,
              fieldNo,
              wireType,
              d,
            );
      }
    }

    return message;
  }

  internalBinaryWrite(
    message: TestEvent,
    writer: IBinaryWriter,
    options: BinaryWriteOptions,
  ): IBinaryWriter {
    /* string test = 1; */
    if (message.test !== '')
      writer.tag(1, WireType.LengthDelimited).string(message.test);

    /* bool isActive = 2; */
    if (message.isActive !== false)
      writer.tag(2, WireType.Varint).bool(message.isActive);

    const u = options.writeUnknownFields;

    if (u !== false)
      (u == true ? UnknownFieldHandler.onWrite : u)(
        this.typeName,
        message,
        writer,
      );

    return writer;
  }
}

/**
 * @generated MessageType for protobuf message test.TestEvent
 */
export const TestEvent = new TestEvent$Type();